node/arbitrage-data-capture/continuous-improvement/requirements.txt
# Elite Continuous Improvement System Requirements

# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Async and Performance
uvloop>=0.17.0
aiofiles>=23.0.0
aioredis>=2.0.0
aiokafka>=0.8.0
asyncpg>=0.28.0
motor>=3.3.0

# Database Clients
clickhouse-driver>=0.2.6
clickhouse-pool>=0.5.3
redis>=5.0.0
psycopg2-binary>=2.9.0

# ML and Deep Learning
torch>=2.0.0
mlflow>=2.8.0
optuna>=3.4.0
xgboost>=2.0.0
lightgbm>=4.1.0
prophet>=1.1.5
river>=0.19.0

# Reinforcement Learning
gym>=0.26.0
stable-baselines3>=2.1.0

# Data Processing
msgpack>=1.0.0
lz4>=4.3.0
xxhash>=3.4.0
pyarrow>=14.0.0

# Monitoring and Profiling
prometheus-client>=0.19.0
py-spy>=0.3.0
psutil>=5.9.0
GPUtil>=1.4.0
memory-profiler>=0.61.0

# Observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-asyncio>=0.42b0
opentelemetry-exporter-jaeger>=1.21.0

# Caching
aiocache>=0.12.0

# Visualization
plotly>=5.18.0
kaleido>=0.2.1

# Kubernetes
kubernetes>=28.0.0

# Docker
docker>=6.1.0

# Configuration
pyyaml>=6.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0